Course fix & classes finished

+++ database/migrations/2016_05_15_175617_create_classes_table.php
@@ -14,7 +14,7 @@ class CreateClassesTable extends Migration
     {
         Schema::create('classes', function(Blueprint $table) {
             $table->increments('id');
-            $table->tinyInteger('year');
+            $table->mediumInteger('year');
             $table->string('label', 3);
         });
     }
